Extract atom-style feed script into its own package
The Atom feed's human-readable styling script (from rss.style) and its
stylesheet previously lived loose under apps/site (a 204-line .js asset
imported `?url`, plus water.light.css and RssStyleFavicon.svg in
public/static/feed). Move them into a dedicated pkgroll-built workspace
package, @adrieldev/atom-style, matching how esm-wrapper/cli are structured.

- packages/atom-style: TypeScript source built to dist/atom-style.js, imported
  by apps/site/src/pages/atom.xml.ts as `@adrieldev/atom-style?url` (the built,
  transpiled JS — a `.ts?url` import would emit raw TypeScript).
- water.light.css moves into the package, exported as a subpath and imported
  `?url` by the feed, passed to the script via a `data-water-css` attribute.
- The RSS.style logo is inlined in the script as a `data:` URI (pkgroll has no
  asset loader for .svg/.txt/?raw), so RssStyleFavicon.svg is no longer needed.
- Wire build ordering (`^build` + predev) and drop the now-dead static/feed
  ignore entries from .prettierignore and the shared ESLint config.
