+++ .prettierignore
@@ -1,5 +1,5 @@
 node_modules
 dist
 .astro
-static/feed
+packages/atom-style/assets
