+++ apps/site/package.json
@@ -16,6 +16,7 @@
     "format:fix": "prettier --write ."
   },
   "dependencies": {
+    "@adrieldev/atom-style": "workspace:*",
     "@adrieldev/esm-wrapper": "workspace:*",
     "@docsearch/css": "4.6.3",
     "@fontsource-variable/jetbrains-mono": "5.2.8",
