+++ packages/eslint-configs/index.js
@@ -13,7 +13,7 @@ import globals from 'globals'
 import tseslint from 'typescript-eslint'
 
 export default defineConfig([
-  globalIgnores(['**/dist/*', '**/.astro/*', '**/static/feed/*']),
+  globalIgnores(['**/dist/*', '**/.astro/*']),
   { files: ['**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'], plugins: { js }, extends: ['js/recommended'] },
   {
     files: ['**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
